API Index

EnvironmentalTransport.AdvectionOperator

EnvironmentalTransport.BCArray

EnvironmentalTransport.ZeroGradBC

EnvironmentalTransport.ZeroGradBCArray

EnvironmentalTransport.Puff

EnvironmentalTransport.Sofiev2012PlumeRise

EnvironmentalTransport.get_vf

EnvironmentalTransport.get_Δ

EnvironmentalTransport.stencil_size

EnvironmentalTransport.upwind1_stencil

EnvironmentalTransport.vf_x

EnvironmentalTransport.vf_y

EnvironmentalTransport.vf_z

EnvironmentalTransport.Δf

API Documentation

EnvironmentalTransport.AdvectionOperator — Type

Create an EarthSciMLBase.Operator that performs advection. Advection is performed using the given stencil operator (e.g. l94_stencil or ppm_stencil). p is an optional parameter set to be used by the stencil operator. bc_type is the boundary condition type, e.g. ZeroGradBC().

Wind field data will be added in automatically if available. Currently the only valid source of wind data is EarthSciData.GEOSFP.

EnvironmentalTransport.BCArray — Type

An array with external indexing implemented for boundary conditions.

EnvironmentalTransport.ZeroGradBC — Type

Zero gradient boundary conditions.

EnvironmentalTransport.ZeroGradBCArray — Type

An array with zero gradient boundary conditions.

EnvironmentalTransport.Puff — Method

Puff(
    di::EarthSciMLBase.DomainInfo;
    buffer_cells,
    name
) -> ModelingToolkit.ODESystem

Create a Lagrangian transport model which advects a "puff" or particle of matter within a fluid velocity field.

Model boundaries are set by the DomainInfo argument. The model sets boundaries at the ground and model bottom and top, preventing the puff from crossing those boundaries. If the puff reaches one of the horizontal boundaries, the simulation is stopped.

Keyword arguments

buffer_cells: The distance (expressed in a number of DomainInfo grid cells) to use as a buffer around the horizontal edge of the domain to avoid data loader interpolation errors. The effective size of the domain will be reduce by 2× this amount (default = 1)

EnvironmentalTransport.Sofiev2012PlumeRise — Method

Wildfire plume rise model based on Sofiev et al. (2012) [1].

[1] Sofiev, M., Ermakova, T., and Vankevich, R.: Evaluation of the smoke-injection height from wild-land fires using remote-sensing data, Atmos. Chem. Phys., 12, 1995–2006, https://doi.org/10.5194/acp-12-1995-2012, 2012.

EnvironmentalTransport.get_vf — Method

get_vf(domain, varname, data_f)

Return a function that gets the wind velocity at a given place and time for the given varname. data_f should be a function that takes a time and three spatial coordinates and returns the value of the wind speed in the direction indicated by varname.

EnvironmentalTransport.get_Δ — Method

get_Δ(domain, tff, pvaridx)

Return a function that gets the grid spacing at a given place and time for the given varname.

EnvironmentalTransport.stencil_size — Method

Return the left and right stencil size of the first-order upwind stencil.

EnvironmentalTransport.upwind1_stencil — Method

upwind1_stencil(ϕ, U, Δt, Δz; p)

First-order upwind advection in 1-D: https://en.wikipedia.org/wiki/Upwind_scheme.

ϕ is the scalar field at the current time step, it should be a vector of length 3 (1 cell on the left, the central cell, and 1 cell on the right).

U is the velocity at both edges of the central grid cell, it should be a vector of length 2.

Δt is the length of the time step.

Δz is the grid spacing.

The output will be time derivative of the central index (i.e. index 2) of the ϕ vector (i.e. dϕ/dt).

Δt and p are not used, but are function arguments for consistency with other operators.

EnvironmentalTransport.vf_x — Method

Get a value from the x-direction velocity field.

EnvironmentalTransport.vf_y — Method

Get a value from the y-direction velocity field.

EnvironmentalTransport.vf_z — Method

Get a value from the z-direction velocity field.

EnvironmentalTransport.Δf — Method

function to get grid deltas.