API Index

EnvironmentalTransport.AdvectionOperator

EarthSciMLBase.couple

EnvironmentalTransport.advect_1d_op

EnvironmentalTransport.get_vf

EnvironmentalTransport.get_Δ

EnvironmentalTransport.l94_stencil

EnvironmentalTransport.orderby_op

EnvironmentalTransport.ppm_stencil

EnvironmentalTransport.simulator_advection_1d

EnvironmentalTransport.stencil_size

EnvironmentalTransport.stencil_size

EnvironmentalTransport.tensor_advection_op

EnvironmentalTransport.zerograd_bc_op

API Documentation

EnvironmentalTransport.AdvectionOperator — Type

Create an EarthSciMLBase.Operator that performs advection. Advection is performed using the given stencil operator (e.g. l94_stencil or ppm_stencil). p is an optional parameter set to be used by the stencil operator.

EarthSciMLBase.couple — Method

couple(c, op)

Couple the advection operator into the CoupledSystem. This function mutates the operator to add the windfield variables. There must already be a source of wind data in the coupled system for this to work. Currently the only valid source of wind data is EarthSciData.GEOSFP.

EnvironmentalTransport.advect_1d_op — Method

advect_1d_op(dtype, shape, stencil, v_f, Δx_f, Δt; p)

Create a SciMLOperator to perform 1D advection.

Arguments: * dtype: The data type of the input and output arrays, e.g. Float64 or Float32. * shape: The shape of the input vector or matrix, e.g. (8,) or (8,10). If the input is a matrix, 1D advection will be applied to each of the columns in the matrix. * stencil: The stencil operator, e.g. l94_stencil or ppm_stencil. * v_f: A function to get the wind velocity at a given place and time. For vector inputs the function signature should be v_f(i, t), where i is the staggered-grid index and t is time. For matrix, inputs, the signature should be v_f(i,j,t), where i is a staggered grid index, j is the non-staggered grid column index (where max(i) == shape(2)), andtis time. *Δxf: A function to get the grid spacing at a given place and time. For vector inputs the function signature should beΔxf(i, t), whereiis the grid index andtis time. For matrix, inputs, the signature should bev_f(i,j,t), whereiis a grid index,jis the column index (wheremax(i) == shape(2)), and t is time. * Δt: The time step size, which is assumed to be fixed. * p: Optional parameters to pass to the stencil function.

EnvironmentalTransport.get_vf — Method

get_vf(sim, varname, data_f, idx_f)

Return a function that gets the wind velocity at a given place and time for the given varname. vardict should be a dictionary with keys that are strings with the possible varnames and values that are the corresponding variables in the ODESystem that should be used to get the wind velocity values. idx_f should be an index function of the type returned by EnvironmentalTransport.orderby_op. data_f should be a function that takes a time and three spatial coordinates and returns the value of the wind speed in the direction indicated by varname.

EnvironmentalTransport.get_Δ — Method

get_Δ(sim, varname)

Return a function that gets the grid spacing at a given place and time for the given varname.

EnvironmentalTransport.l94_stencil — Method

l94_stencil(ϕ, U, Δt, Δz; kwargs...)

L94 advection in 1-D (Lin et al., 1994)

ϕ is the scalar field at the current time step, it should be a vector of length 5.

U is the velocity at both edges of the central grid cell, it should be a vector of length 2.

Δt is the length of the time step.

Δz is the grid spacing.

The output will be time derivative of the central index (i.e. index 3) of the ϕ vector (i.e. dϕ/dt).

(The output is dependent on the Courant number, which depends on Δt, so Δt needs to be an input to the function.)

EnvironmentalTransport.orderby_op — Method

orderby_op(dtype, shape, index; p)

Create a SciMLOperator to reorder a tensor into a matrix, where each column in the matrix is an element in the given index of the original tensor. dtype is the type of the input and output data, e.g. Float64 or Float32, and shape is the shape of the tensor, e.g. [2,3,5].

The inverse of the operator can be used to put the tensor back into the orinal order. So if op is the operator and u is the original tensor, then: x = op * u[:] returns the reordered tensor x, and u_prime = inv(op) * x returns a vector version of the orginal tensor, so uprime == u[:].

This function returns the operator and also a function idx_f that takes a column number of the transformed matrix as an input and returns a vector of CartesianIndexes in the original tensfor that make up that transformed matrix column.

EnvironmentalTransport.ppm_stencil — Method

ppm_stencil(ϕ, U, Δt, Δz; kwargs...)

PPM advection in 1-D (Collela and Woodward, 1984)

ϕ is the scalar field at the current time step, it should be a vector of length 8 (3 cells on the left, the central cell, and 4 cells on the right).

U is the velocity at both edges of the central grid cell, it should be a vector of length 2.

Δt is the length of the time step.

Δz is the grid spacing.

The output will be time derivative of the central index (i.e. index 4) of the ϕ vector (i.e. dϕ/dt).

(The output is dependent on the Courant number, which depends on Δt, so Δt needs to be an input to the function.)

EnvironmentalTransport.simulator_advection_1d — Method

simulator_advection_1d(sim, op, varname; p)

Create a 1D advection SciMLOperator for the given variable name varname. vardict should be a dictionary with keys that are strings with the possible varnames and values that are the corresponding variables in the ODESystem that should be used to get the wind velocity values. p is an optional parameter set that can be passed to the stencil function.

EnvironmentalTransport.stencil_size — Method

Return the left and right stencil size of the L94 stencil.

EnvironmentalTransport.stencil_size — Method

Return the left and right stencil size of the PPM stencil.

EnvironmentalTransport.tensor_advection_op — Method

tensor_advection_op(
    dtype,
    shape,
    index,
    stencil;
    bc_opf,
    p,
    kwargs...
)

Return a SciMLOperator that performs 1D advection on the dimension of a tensor given by index, where the original tensor has the given data type dtype (e.g. Float32 or Float64), and given shape (e.g. (10, 20, 30)). Advection is performed using the given stencil operator (e.g. l94_stencil or ppm_stencil).

Optionally, a function can be specified to create boundary conditions, where the function should have the signature bc_opf(vector_length, stencil). See the default boundary condition operator EnvironmentalTransport.zerograd_bc_op for more information.

This function returns a function that creates the the operator and also a function idx_f that takes a column number of the transformed matrix as an input and returns a vector of CartesianIndexes in the original tensor that make up that transformed matrix column.

The first returned function (the function that creates the operator) can be called with the following arguments to create the operator:

*  `v_f`: A function to get the wind velocity at a given place and time. For vector inputs
        the function signature should be `v_f(i, t)`, where `i` is the staggered-grid 
        index and `t` is time. For matrix, inputs, the signature should be `v_f(i,j,t)`,
        where `i` is a staggered grid index, `j` is the non-staggered grid column index
        (where `max(i) == shape(2)`)`, and `t` is time.
* `Δx_f`: A function to get the grid spacing at a given place and time. For vector inputs
        the function signature should be `Δx_f(i, t)`, where `i` is the grid 
        index and `t` is time. For matrix, inputs, the signature should be `v_f(i,j,t)`,
        where `i` is a grid index, `j` is the column index
        (where `max(i) == shape(2)`)`, and `t` is time.
* `Δt`: The time step size, which is assumed to be fixed.
* `p`: Optional parameters to pass to the stencil function.

The reason for the two-step process is that Δv_f and Δx_fmay be dependent onidxf, so we need to returnidxf` before we create the operator.

EnvironmentalTransport.zerograd_bc_op — Method

zerograd_bc_op(vec_length, stencil)

Create an SciMLOperator which adds zero gradient Newmann boundary conditions to a 1D array of length vec_length, which work with the given stencil.